refactor: 클래스명 및 위치 변경에 따른 동기화
