+++ src/main/java/umc/stockoneqback/global/enumconfig/EnumStandard.java
@@ -1,4 +1,4 @@
-package umc.stockoneqback.global;
+package umc.stockoneqback.global.enumconfig;
 
 public interface EnumStandard {
