+++ src/main/java/umc/stockoneqback/user/exception/UserErrorCode.java
@@ -3,7 +3,7 @@ package umc.stockoneqback.user.exception;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
-import umc.stockoneqback.global.exception.ErrorCode;
+import umc.stockoneqback.global.base.ErrorCode;
 
 @Getter
 @RequiredArgsConstructor
